% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table.R
\name{life.table}
\alias{life.table}
\title{Life Table Function}
\usage{
life.table(mx, sex = c("male", "female"), ages = names(mx),
  radix = 1, open.age = 130)
}
\arguments{
\item{mx}{Vector of age-specific mortality rates nmx. The elements correspond to 1m0, 4m1, 5m5, 5m10, \dots. 
It can have no more than 28 elements which corresponds to age up to 130.}

\item{sex}{For which sex is the life table.}

\item{ages}{Defines the age groups of \code{mx}. If \code{mx} has no names, the \code{ages} corresponds to 0, 1, 5, 10, \dots.}

\item{radix}{Base of the life table.}

\item{open.age}{Open age group. If smaller than the last age group of \code{mx}, the life table is truncated.}
}
\description{
Function for obtaining life table quantities from mortality rates.
}
\details{
Computes a life table corresponding to given mortality rates for 5-years age groups.
}
\examples{
data(mxF, e0Fproj, package = "wpp2017")
# get female mortality of Mexico for the current year
country <- "Mexico"
mxf <- subset(mxF, name == country)[,"2010-2015"]
names(mxf) <- c(0,1, seq(5, 100, by=5))
life.table(mxf, sex = "female")

}
