% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortCast-data.R
\docType{data}
\name{LQcoef}
\alias{LQcoef}
\title{Coefficients for the Log-Quadratic Mortality Model}
\format{
Data frame containing columns \dQuote{sex}, \dQuote{age}, \dQuote{ax}, \dQuote{bx}, \dQuote{cx}, \dQuote{vx}. 
    Rows correspond to sex and age groups.
}
\source{
\url{http://www.demog.berkeley.edu/~jrw/LogQuad}
}
\usage{
data(LQcoef)
}
\description{
Data object containing a table of coefficients to be used in the Log-Quadratic Model 
    as implemented in the \code{\link{logquad}} function.
}
\examples{
data(LQcoef)
head(LQcoef)


}
\references{
Wilmoth, J., Zureick, S., Canudas-Romo, V., Inoue, M., Sawyer, C. (2012). 
     A Flexible Two-Dimensional Mortality Model for Use in Indirect Estimation. 
     Population studies, 66(1), 1-28. \doi{doi:10.1080/00324728.2011.611411}
}
\seealso{
\code{\link{logquad}}
}
\keyword{datasets}
