% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortCast-data.R
\docType{data}
\name{RhoFemales}
\alias{RhoFemales}
\alias{rhoPMD}
\alias{RhoMales}
\title{Pattern Mortality Decline Lookup Table}
\format{An object of class ; see .}
\usage{
data(rhoPMD)

RhoMales
}
\description{
Pattern Mortality Decline Lookup Table
}
\keyword{datasets}
