% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortCast-data.R
\docType{data}
\name{MLTlookup}
\alias{MLTlookup}
\title{Model Life Tables Lookup Table}
\format{An object of class .}
\usage{
data(MLTlookup)
}
\description{
Model Life Tables Lookup Table
}
\keyword{datasets}
