% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{mortcast.blend}
\alias{mortcast.blend}
\title{Mortality Prediction by Method Blending}
\usage{
mortcast.blend(e0m, e0f, meth1 = "lc", meth2 = "mlt", weights = c(1,
  0.5), apply.kannisto = TRUE, min.age.groups = 28,
  meth1.args = NULL, meth2.args = NULL, kannisto.args = NULL)
}
\arguments{
\item{e0m}{A time series of future male life expectancy.}

\item{e0f}{A time series of future female life expectancy.}

\item{meth1}{Character string giving the name of the first method to blend. It is one of 
\dQuote{lc}, \dQuote{pmd} or \dQuote{mlt}, corresponding to Coherent Lee-Carter (function \code{\link{mortcast}}), 
 Pattern Mortality Decline (function \code{\link{copmd}}) and 
 Model Life Tables (function \code{\link{mltj}}), respectively.}

\item{meth2}{Character string giving the name of the second method to blend. 
One of the same choices as \code{meth1}.}

\item{weights}{Numeric vector with values between 0 and 1 giving the weight of \code{meth1}.
If it is a single value, the same weight is applied for all time periods. 
If it is a vector of size two, it is assumed these are weights for the first and the last
time period. Remaining weights will be interpolated. Note that \code{meth2} is weighted 
by \code{1 - weights}.}

\item{apply.kannisto}{Logical. If \code{TRUE} and if any of the methods results in less than 
\code{min.age.groups} age categories, the coherent Kannisto method (\code{\link{cokannisto}}) 
is applied to extend the age groups into old ages.}

\item{min.age.groups}{Minimum number of age groups. Triggers the application of Kannisto, see above.}

\item{meth1.args}{List of arguments passed to the function that corresponds to \code{meth1}.}

\item{meth2.args}{List of arguments passed to the function that corresponds to \code{meth2}.}

\item{kannisto.args}{List of arguments passed to the \code{\link{cokannisto}} function if Kannisto is applied.}
}
\value{
List with elements \code{female} and \code{male}, each of which contains a matrix \code{mx}
    with the predicted mortality rates. In addition, it contains elements \code{meth1res} and \code{meth2res}
    which contain the results of the functions corresponding to the two methods. 
    Elements \code{meth1} and \code{meth2} contain the names of the methods. 
    A vector \code{weights} contains the final (possibly interpolated) weights.
}
\description{
Predict age-specific mortality rates using a blend of two different methods (Coherent Lee-Carter, 
    Pattern Mortality Decline, or Model Life Tables). Weights can be applied to fine-tune the blending mix.
}
\details{
The function allows to combine two different methods using given weights.
    The weights can change over time - by default they are interpolated from the starting weight 
    to the end weight. The projection is done for both sexes, so that coherent methods can be applied.
}
\examples{
data(mxM, mxF, e0Fproj, e0Mproj, package = "wpp2017")
country <- "Brazil"
# estimate parameters from historical mortality data
mxm <- subset(mxM, name == country)[,4:16]
mxf <- subset(mxF, name == country)[,4:16]
rownames(mxm) <- rownames(mxf) <- c(0,1, seq(5, 100, by=5))
lcest <- lileecarter.estimate(mxm, mxf)
# project into future
e0f <- as.numeric(subset(e0Fproj, name == country)[-(1:2)])
e0m <- as.numeric(subset(e0Mproj, name == country)[-(1:2)])
# Blend LC and MLT
pred1 <- mortcast.blend(e0m, e0f, meth1 = "lc", meth2 = "mlt",
    meth1.args = list(lc.pars = lcest),
    meth2.args = list(type = "CD_North"),
    weights = c(1,0.25))
# Blend PMD and MLT
pred2 <- mortcast.blend(e0m, e0f, meth1 = "pmd", meth2 = "mlt",
    meth1.args = list(mxm0 = mxm[, "2010-2015"],
                      mxf0 = mxf[, "2010-2015"]))
# plot projection by time
plotmx <- function(pred, iage, main) 
    with(pred, {
        # blended projections 
        plot(female$mx[iage,], type="l", 
            ylim=range(meth1res$female$mx[iage,], 
                       meth2res$female$mx[iage,]),
            ylab="female mx", xlab="Time", main=main, col = "red")
        lines(meth1res$female$mx[iage,], lty = 2)
        lines(meth2res$female$mx[iage,], lty = 3)
        legend("topright", legend=c("blend", meth1, meth2),
               lty = 1:3, col = c("red", "black", "black"), bty = "n")
    })
age.group <- 3 # 5-9 years old
par(mfrow=c(1,2))
plotmx(pred1, age.group, "LC-MLT (age 5-9)")
plotmx(pred2, age.group, "PMD-MLT (age 5-9)")

}
\seealso{
\code{\link{mortcast}}, \code{\link{copmd}}, \code{\link{mltj}}, 
    \code{\link{cokannisto}}
}
