% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_methods.R
\name{pmd}
\alias{pmd}
\alias{pmdj}
\title{Pattern of Mortality Decline Prediction}
\usage{
pmd(e0, mx0, sex = c("male", "female"), interp.rho = FALSE,
  kranges = c(0.01, 25), keep.lt = FALSE, keep.rho = FALSE)

pmdj(e0m, e0f, mxm0, mxf0, ...)
}
\arguments{
\item{e0}{A time series of target life expectancy.}

\item{mx0}{A vector with starting age-specific mortality rates.}

\item{sex}{Either "male" or "female".}

\item{interp.rho}{Logical controlling if the \eqn{\rho} coefficients should be interpolated 
(\code{TRUE}) or binned (\code{FALSE}).}

\item{kranges}{A vector of size two, giving the min and max of the \eqn{k} parameter which is 
estimated to match the the target \code{e0}.}

\item{keep.lt}{Logical. If \code{TRUE} additional life table columns are kept in the 
resulting object.}

\item{keep.rho}{Logical. If \code{TRUE} the \eqn{\rho} coefficients are included in the resulting object.}

\item{e0m}{A time series of target male life expectancy.}

\item{e0f}{A time series of target female life expectancy.}

\item{mxm0}{A vector with starting age-specific male mortality rates.}

\item{mxf0}{A vector with starting age-specific female mortality rates.}

\item{\dots}{Additional arguments passed to the underlying function.}
}
\value{
List with elements a matrix \code{mx}
    with the predicted mortality rates. If \code{keep.lt} is \code{TRUE}, it also 
    contains matrices \code{sr} (survival rates), and life table quantities \code{Lx} and \code{lx}.
    If \code{keep.rho} is \code{TRUE}, it contains a matrix \code{rho} where columns correpond 
    to the values in the \code{e0} vector and rows correspond to age groups.
}
\description{
Predict age-specific mortality rates using the Pattern of mortality decline (PMD) method.
}
\details{
These functions implements the PMD method published in Andreev et al. (2013). 
    It assumes that the future decline in age-specific mortality will follow a certain pattern 
    with the increase in life expectancy at birth. 
    
    Function \code{pmd} evaluates the method for a single sex, while  \code{pmdj} does it
    for both sexes.
}
\examples{
data(mxF, e0Fproj, package = "wpp2017")
country <- "Hungary"
# get initial mortality for the current year
mxf <- subset(mxF, name == country)[,"2010-2015"]
names(mxf) <- c(0,1, seq(5, 100, by=5))
# get target e0
e0f <- as.numeric(subset(e0Fproj, name == country)[-(1:2)])
# project into future
pred <- pmd(mxf, e0f, sex = "female")
# plot first projection in black and the remaining ones in grey 
plot(pred$mx[,1], type="l", log="y", ylim=range(pred$mx),
    ylab="female mx", xlab="Age", main=country)
for(i in 2:ncol(pred$mx)) lines(pred$mx[,i], col="grey")

}
\references{
Andreev, K. Gu, D., Gerland, P. (2013). Age Patterns of Mortality Improvement by Level of Life Expectancy at Birth with Applications to Mortality Projections. Paper presented at the Annual Meeting
of the Population Association of America, New Orleans, LA. \url{http://paa2013.princeton.edu/papers/132554}.
}
\seealso{
\code{\link{mortcast}}
}
